processing/requirements.txt
appdirs==1.4.3
appnope==0.1.0
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.9.0
bleach==3.1.4
bokeh==2.0.1
boto3==1.11.9
botocore==1.14.9
bs4==0.0.1
certifi==2020.4.5.1
cftime==1.1.1.2
chardet==3.0.4
click==7.1.1
click-plugins==1.1.1
cligj==0.5.0
cloudpickle==1.3.0
cssselect==1.1.0
cycler==0.10.0
Cython==0.29.16
dask==2.14.0
decorator==4.4.2
defusedxml==0.6.0
deprecation==2.0.7
distributed==2.14.0
docutils==0.15.2
entrypoints==0.3
fake-useragent==0.1.11
Fiona==1.8.13.post1
fsspec==0.7.1
geojson==2.5.0
geopandas==0.7.0
h5netcdf==0.8.0
h5py==2.10.0
HeapDict==1.0.1
idna==2.9
importlib-metadata==1.6.0
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.16.0
Jinja2==2.11.1
jmespath==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.2
jupyter-core==4.6.3
kiwisolver==1.2.0
llvmlite==0.31.0
locket==0.2.0
lxml==4.5.0
MarkupSafe==1.1.1
matplotlib==3.2.1
mistune==0.8.4
msgpack==1.0.0
munch==2.5.0
nbconvert==5.6.1
nbformat==5.0.5
netCDF4==1.5.3
notebook==6.0.3
numba==0.48.0
numpy==1.18.2
packaging==20.3
pandas==1.0.3
pandocfilters==1.4.2
parse==1.15.0
parso==0.6.2
partd==1.1.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.1
prometheus-client==0.7.1
prompt-toolkit==3.0.5
psutil==5.7.0
ptyprocess==0.6.0
pyee==7.0.1
Pygments==2.6.1
pymongo==3.10.1
pyparsing==2.4.7
pyppeteer==0.0.25
pyproj==2.6.0
pyquery==1.4.1
pyrsistent==0.16.0
python-dateutil==2.8.1
python-dotenv==0.12.0
pytz==2019.3
PyYAML==5.3.1
pyzmq==19.0.0
redis==3.4.1
requests==2.23.0
requests-html==0.10.0
s3transfer==0.3.3
scipy==1.4.1
Send2Trash==1.5.0
Shapely==1.7.0
six==1.12.0
sortedcontainers==2.1.0
soupsieve==2.0
swifter==0.302
tblib==1.6.0
terminado==0.8.3
testpath==0.4.4
toolz==0.10.0
tornado==6.0.4
tqdm==4.45.0
traitlets==4.3.3
typing-extensions==3.7.4.2
urllib3==1.25.8
w3lib==1.21.0
wcwidth==0.1.9
webencodings==0.5.1
websockets==8.1
widgetsnbextension==3.5.1
wradlib==1.6.2
xarray==0.15.1
xmltodict==0.12.0
zict==2.0.0
zipp==3.1.0
